Avoid qualifying generated enum default values (#98)
* Add regression test for qualified enum name issue

When generating default values for enum fields and the generated code
contains a type with the same name as the namespace, the resulting
qualified enum type name will refer to the class and fail resolution.

* Avoid qualifying generated enum default values

This is a workaround to avoid qualifying the generated enum value
identifier. We know the simple version must never clash in generated
code since it is used in property type declarations.